refactor: move chunking utilities to pkg/common/envoy/request
